+++ src/sentry/javascript/javascript_span.h
@@ -26,6 +26,8 @@ public:
 	virtual void set_attribute(const String &p_key, const Variant &p_value) override;
 	virtual void set_status(SpanStatus p_status) override;
 	virtual void end() override;
+
+	virtual PackedStringArray get_trace_headers() override;
 };
 
 } //namespace sentry::javascript