+++ src/sentry/sentry_options.cpp
@@ -6,6 +6,7 @@
 
 #include <godot_cpp/classes/os.hpp>
 #include <godot_cpp/classes/project_settings.hpp>
+#include <godot_cpp/classes/reg_ex.hpp>
 
 namespace {
 
@@ -129,6 +130,9 @@ void SentryOptions::_define_project_settings(const Ref<SentryOptions> &p_options
 	_define_setting(PropertyInfo(Variant::FLOAT, "sentry/options/sample_rate", PROPERTY_HINT_RANGE, "0.0,1.0"), p_options->sample_rate, false);
 	_define_setting(PropertyInfo(Variant::FLOAT, "sentry/options/traces_sample_rate", PROPERTY_HINT_RANGE, "0.0,1.0"), p_options->traces_sample_rate, false);
 	_define_setting(PropertyInfo(Variant::INT, "sentry/options/trace_lifecycle", PROPERTY_HINT_ENUM, "Static,Stream"), p_options->trace_lifecycle, false);
+	_define_setting("sentry/options/trace_propagation_targets", PackedStringArray(p_options->trace_propagation_targets), false);
+	_define_setting("sentry/options/propagate_traceparent", p_options->propagate_traceparent, false);
+	_define_setting("sentry/options/org_id", p_options->org_id, false);
 	_define_setting(PropertyInfo(Variant::INT, "sentry/options/max_breadcrumbs", PROPERTY_HINT_RANGE, "0, 500"), p_options->max_breadcrumbs, false);
 	_define_setting(PropertyInfo(Variant::INT, "sentry/options/shutdown_timeout_ms", PROPERTY_HINT_RANGE, "0,30000"), p_options->shutdown_timeout_ms, false);
 	_define_setting("sentry/options/send_default_pii", p_options->send_default_pii);
@@ -220,6 +224,10 @@ void SentryOptions::_load_project_settings(const Ref<SentryOptions> &p_options) 
 	p_options->sample_rate = ProjectSettings::get_singleton()->get_setting("sentry/options/sample_rate", p_options->sample_rate);
 	p_options->traces_sample_rate = ProjectSettings::get_singleton()->get_setting("sentry/options/traces_sample_rate", p_options->traces_sample_rate);
 	p_options->trace_lifecycle = (SentryOptions::TraceLifecycle)(int)ProjectSettings::get_singleton()->get_setting("sentry/options/trace_lifecycle", p_options->trace_lifecycle);
+	const PackedStringArray trace_propagation_targets = ProjectSettings::get_singleton()->get_setting("sentry/options/trace_propagation_targets", PackedStringArray(p_options->trace_propagation_targets));
+	p_options->set_trace_propagation_targets(Array(trace_propagation_targets));
+	p_options->propagate_traceparent = ProjectSettings::get_singleton()->get_setting("sentry/options/propagate_traceparent", p_options->propagate_traceparent);
+	p_options->org_id = ProjectSettings::get_singleton()->get_setting("sentry/options/org_id", p_options->org_id);
 	p_options->max_breadcrumbs = ProjectSettings::get_singleton()->get_setting("sentry/options/max_breadcrumbs", p_options->max_breadcrumbs);
 	p_options->shutdown_timeout_ms = ProjectSettings::get_singleton()->get_setting("sentry/options/shutdown_timeout_ms", p_options->shutdown_timeout_ms);
 	p_options->send_default_pii = ProjectSettings::get_singleton()->get_setting("sentry/options/send_default_pii", p_options->send_default_pii);
@@ -394,6 +402,27 @@ void SentryOptions::release_callables() {
 	before_capture_screenshot = Callable();
 }
 
+void SentryOptions::set_trace_propagation_targets(const Array &p_targets) {
+	Array valid_targets;
+	for (const Variant &target : p_targets) {
+		if (target.get_type() == Variant::STRING) {
+			const String text = target;
+			if (!text.is_empty()) {
+				valid_targets.append(target);
+			}
+		} else {
+			ERR_CONTINUE_MSG(target.get_type() != Variant::OBJECT, "Sentry: trace_propagation_targets entries must be String or RegEx values.");
+			Object *object = target;
+			RegEx *regex = Object::cast_to<RegEx>(object);
+			ERR_CONTINUE_MSG(regex == nullptr, "Sentry: trace_propagation_targets entries must be String or RegEx values.");
+			ERR_CONTINUE_MSG(!regex->is_valid(), "Sentry: trace_propagation_targets entries must contain valid regular expressions.");
+			valid_targets.append(target);
+		}
+	}
+
+	trace_propagation_targets = valid_targets;
+}
+
 void SentryOptions::_bind_methods() {
 	BIND_PROPERTY(SentryOptions, PropertyInfo(Variant::STRING, "dsn"), set_dsn, get_dsn);
 	BIND_PROPERTY(SentryOptions, PropertyInfo(Variant::STRING, "release"), set_release, get_release);
@@ -404,6 +433,9 @@ void SentryOptions::_bind_methods() {
 	BIND_PROPERTY(SentryOptions, PropertyInfo(Variant::FLOAT, "sample_rate"), set_sample_rate, get_sample_rate);
 	BIND_PROPERTY(SentryOptions, PropertyInfo(Variant::FLOAT, "traces_sample_rate"), set_traces_sample_rate, get_traces_sample_rate);
 	BIND_PROPERTY(SentryOptions, PropertyInfo(Variant::INT, "trace_lifecycle", PROPERTY_HINT_ENUM, "Static,Stream"), set_trace_lifecycle, get_trace_lifecycle);
+	BIND_PROPERTY(SentryOptions, PropertyInfo(Variant::ARRAY, "trace_propagation_targets"), set_trace_propagation_targets, get_trace_propagation_targets);
+	BIND_PROPERTY(SentryOptions, PropertyInfo(Variant::BOOL, "propagate_traceparent"), set_propagate_traceparent, is_propagate_traceparent_enabled);
+	BIND_PROPERTY(SentryOptions, PropertyInfo(Variant::STRING, "org_id"), set_org_id, get_org_id);
 	BIND_PROPERTY(SentryOptions, PropertyInfo(Variant::INT, "max_breadcrumbs"), set_max_breadcrumbs, get_max_breadcrumbs);
 	BIND_PROPERTY(SentryOptions, PropertyInfo(Variant::INT, "shutdown_timeout_ms", PROPERTY_HINT_RANGE, "0,30000"), set_shutdown_timeout_ms, get_shutdown_timeout_ms);
 	BIND_PROPERTY(SentryOptions, PropertyInfo(Variant::BOOL, "send_default_pii"), set_send_default_pii, is_send_default_pii_enabled);